requirements.txt
numpy==1.24.3
opencv_python==4.11.0.86
Pillow==11.3.0
scipy==1.16.1
torch==2.4.1
torchvision==0.19.1
